Updater: Change the update jitter to a double instead of int.
Bug: 1144151
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2640732
Commit-Queue: Mila Green <milagreen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#845437}

+++ chrome/updater/constants.h
@@ -209,7 +209,7 @@ constexpr int kUninstallPingReasonUserNotAnOwner = 1;
 // The file downloaded to a temporary location could not be moved.
 constexpr int kErrorFailedToMoveDownloadedFile = 5;
 
-constexpr int kInitialDelay = 60;
+constexpr double kInitialDelay = 60;
 
 }  // namespace updater
